Split Tree into `buddy_allocator` macro -- will be used for heap
