+++ kernel/src/lib.rs
@@ -30,13 +30,10 @@ extern crate bitflags;
 #[macro_use]
 extern crate lazy_static;
 
-use core::mem;
 use drivers::keyboard::{Keyboard, KeyEventType, Ps2Keyboard};
 use drivers::keyboard::keymap;
 use drivers::ps2;
 use terminal::TerminalOutput;
-use memory::bootstrap_heap;
-use memory::physical_allocator::{PHYSICAL_ALLOCATOR, BLOCKS_IN_TREE, Block};
 
 #[cfg(not(test))]
 mod lang;
@@ -61,24 +58,9 @@ pub extern fn kmain(multiboot_info_addr: usize) -> ! {
 
     // Parse and print mb info
     let mb_info = unsafe { multiboot2::load(multiboot_info_addr) };
-    let memory_map = mb_info.memory_map_tag()
-        .expect("Expected a multiboot2 memory map tag, but it is not present!");
-    print_memory_info(memory_map);
 
-    // Set up bootstrap heap
-    let end_address = mb_info.end_address() as *const u8;
-    let end_address = unsafe { end_address.offset(
-        end_address.align_offset(mem::align_of::<[Block; BLOCKS_IN_TREE]>()) as isize
-    )};
-    let heap_start = end_address;
-    unsafe { bootstrap_heap::BOOTSTRAP_HEAP.init_unchecked(heap_start as usize); }
-
-    // Set up physical frame allocator
-    PHYSICAL_ALLOCATOR.init(1, &[]); // TODO handle holes & # of GiB properly
-
-    for _ in 0..4 {
-        debug!("Allocated {:?}", PHYSICAL_ALLOCATOR.allocate(0).unwrap());
-    }
+    // Init memory
+    memory::init_memory(&mb_info);
 
     // Initialize the PS/2 controller and run the keyboard echo loop
     let mut controller = ps2::CONTROLLER.lock();
@@ -86,7 +68,13 @@ pub extern fn kmain(multiboot_info_addr: usize) -> ! {
         Ok(_) => info!("ps2c: init successful"),
         Err(error) => error!("ps2c: {:?}", error),
     }
+    keyboard_echo_loop(&mut controller);
 
+    halt()
+}
+
+/// Say hello to the user and print flower
+fn say_hello() {
     terminal::STDOUT.write().clear().expect("Screen clear failed");
 
     print_flower().expect("Flower print failed");
@@ -101,14 +89,6 @@ pub extern fn kmain(multiboot_info_addr: usize) -> ! {
     // Reset colors
     terminal::STDOUT.write().set_color(color!(White on Black))
         .expect("Color should be supported");
-
-    let mut controller = ps2::CONTROLLER.lock();
-    match controller.initialize() {
-        Ok(_) => info!("ps2c: init successful"),
-        Err(error) => error!("ps2c: {:?}", error),
-    }
-
-    keyboard_echo_loop(controller)
 }
 
 fn print_flower() -> Result<(), terminal::TerminalOutputError<()>> {
@@ -142,14 +122,6 @@ fn keyboard_echo_loop(controller: &mut ps2::Controller) {
         }
     } else {
         error!("kbd: enable unsuccessful");
-    }
-}
-
-fn print_memory_info(memory_map: &multiboot2::MemoryMapTag) {
-    debug!("mem: Usable memory areas: ");
-    for area in memory_map.memory_areas() {
-        debug!(" 0x{:x} to 0x{:x}",
-                 area.start_address(), area.end_address());
     }
 }
